Add DatabaseConnector interface, and move execute logic to StandardDatabaseConnector
